refactor: change logger source file location
